move lib to src subfolder, remove symlinks
